[MR] BE, FE 통합 Merge

+++ back/src/main/java/com/realive/domain/auction/Bid.java
@@ -1,4 +1,4 @@
-                                package com.realive.domain.auction;
+package com.realive.domain.auction;
 
 import com.realive.domain.common.BaseTimeEntity;
 import jakarta.persistence.*;
@@ -23,13 +23,11 @@ public class Bid extends BaseTimeEntity {
     private Integer auctionId;
 
     @Column(name = "customer_id")
-    private Integer customerId;
+    private Long customerId;
 
     @Column(name = "bid_price")
     private Integer bidPrice;
 
     @Column(name = "bid_time")
     private LocalDateTime bidTime;
-
-
 }
